Moved endpoint to /moves and extracted PackDays function
