+++ swagger/internal.yaml
@@ -3391,6 +3391,39 @@ paths:
             $ref: '#/definitions/MovePayload'
         500:
           description: server error
+  /moves/{moveId}/move_dates_summary:
+    get:
+      summary: Returns projected move-related dates for a given move date
+      description: Returns projected move-related dates for a given move date
+      operationId: showMoveDatesSummary
+      tags:
+        - moves
+      parameters:
+        - in: path
+          name: moveId
+          type: string
+          format: uuid
+          required: true
+          description: UUID of the move
+        - in: query
+          name: move_date
+          type: string
+          format: date
+          required: true
+          description: The chosen move date
+      responses:
+        200:
+          description: List of projected move-related dates
+          schema:
+            $ref: '#/definitions/MoveDatesSummaryPayload'
+        400:
+          description: invalid request
+        401:
+          description: request requires user authentication
+        403:
+          description: user is not authorized
+        500:
+          description: internal server error
   /personally_procured_moves/{personallyProcuredMoveId}/create_ppm_attachments:
     post:
       summary: Creates PPM attachments PDF
@@ -4043,39 +4076,6 @@ paths:
               type: string
               format: date
               example: "2018-09-25"
-        400:
-          description: invalid request
-        401:
-          description: request requires user authentication
-        403:
-          description: user is not authorized
-        500:
-          description: internal server error
-  /calendar/{moveId}/move_dates_summary:
-    get:
-      summary: Returns projected move-related dates for a given move date
-      description: Returns projected move-related dates for a given move date
-      operationId: showMoveDatesSummary
-      tags:
-        - calendar
-      parameters:
-        - in: path
-          name: moveId
-          type: string
-          format: uuid
-          required: true
-          description: UUID of the move
-        - in: query
-          name: move_date
-          type: string
-          format: date
-          required: true
-          description: The chosen move date
-      responses:
-        200:
-          description: List of projected move-related dates
-          schema:
-            $ref: '#/definitions/MoveDatesSummaryPayload'
         400:
           description: invalid request
         401:
